backend/requirements.txt
FastAPI
uvicorn
langchain
langchain-openai
langchain-community
langgraph
pytest
pytest-asyncio
pytest-mock
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
requests
beautifulsoup4
openai
pydantic>=2.7.0
pydantic-settings>=2.2.0
structlog>=24.1.0